Split different methods of getting product stock into different classes, implementing combined interface; introduced factory to decide which one to use at DI-time.
